pyproject.toml
[tool.poetry]
name = "tg-chatbot"
version = "0.1.0"
description = ""
authors = ["niyoh <zhrh120@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = {extras = ["callback-data"], version = "^20.2"}
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
revChatGPT = "*"
EdgeGPT = "*"
structlog = "^22.3.0"
markdown = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
